requirements.txt
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cmake==3.26.3
cryptography==40.0.2
customtkinter==5.1.2
darkdetect==0.8.0
decorator==5.1.1
exceptiongroup==1.1.1
executing==1.2.0
fancycompleter==0.9.1
filelock==3.12.0
h11==0.14.0
huggingface-hub==0.13.4
idna==3.4
ipdb==0.13.13
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
lit==16.0.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.1
nltk==3.4.5
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
outcome==1.2.0
packaging==23.1
parso==0.8.3
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.17
Pygments==2.15.0
pypdf==3.7.1
pypdf2==3.0.1
pyrepl==0.9.0
PySocks==1.7.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
scipy==1.10.1
selenium==4.8.3
sentencepiece==0.1.98
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
sympy==1.11.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
trio==0.22.0
trio-websocket==0.10.2
triton==2.0.0
typing==3.7.4.3
typing-extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
wget==3.2
wmctrl==0.4
wsproto==1.2.0
